split Volume/Muted flags, use standard DBus interface to notify about changes to these props

+++ src/event.rs
@@ -1,5 +1,5 @@
 #[derive(Debug)]
 pub(crate) struct Event {
-    pub(crate) volume_changed: Option<f32>,
+    pub(crate) volume_changed: Option<u32>,
     pub(crate) muted_changed: Option<bool>,
 }
